requirements.txt
bleach==2.1.2
Django==2.0
django-autoslug==1.9.3
django-contrib-postgres==0.0.1
django-markdownify==0.3.0
django-model-search==2.0
django-modeltranslation==0.12.1
gunicorn==19.7.1
html5lib==1.0.1
Markdown==2.6.10
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
webencodings==0.5.1
